DP: refactor parse class and module
