+++ data-pipes/parse/identities.rb
@@ -1,8 +1,15 @@
 #!/usr/bin/sh ruby
 
-require 'csv'
+$LOAD_PATH << 'parse'
 
-class Identities
+load 'parse/parser.rb'
+require 'units_identities'
+require 'csv'
+require 'json'
+
+class Identities < Parser
+
+  include UnitsIdentities
 
   attr_reader :ids
   attr_reader :sections, :headers, :file
@@ -20,24 +27,6 @@ class Identities
     clean
     pass
   end
-
-  # TODO yml source
-  IDENTITY_WORDS = %w{ CID EID EDT ORDER ACCOUNT Med Technician ID NAME Room Loc }
-  IDENTITY_PHRASES = ["Test ind", "Referred by", "Confirmed by", "Vent. rate", "PR interval", "QRS duration", "QT/QTc", "P-R-T axes"]
-
-  IDENTITIES = (
-    IDENTITY_WORDS.sort   {|a,b| a.length <=> b.length }.reverse +
-    IDENTITY_PHRASES.sort {|a,b| a.length <=> b.length }.reverse
-  ).freeze
-
-  UNIT_RATIOS = %w{ mm/s mm/mV } # '|' chars will break regexes
-  UNIT_QUANTITIES = %w{ lb in YR BPM ms Hz }
-  SEX = %w{ Male Female }
-  RACE = [ "Caucasian", "Latino", "African", "Asian", "Native America", "Decline to self identify"]
-  UNITS = (UNIT_RATIOS + UNIT_QUANTITIES + SEX + RACE).freeze
-
-  DATE_INPUT_FORMAT  = /[2-9]{2}-[A-Z]{3}-[0-9]{4}/.freeze
-  DATE_OUTPUT_FORMAT = "%F %H:%M:%S"
 
   def parse
     return @sections unless @sections.empty?
@@ -136,34 +125,4 @@ class Identities
     end
   end
 
-  # TODO for a superclass method
-  def pass save_target=:disk
-    send save_target
-    puts "saved to #{save_target}"
-  end
-
-  def disk
-    CSV.open(filename_suffixed+'.csv', 'wb') do |csv|
-      csv << @headers
-      @sections.each {|s| csv << s }
-    end
-  end
-
-  def filename_suffixed
-    @filename.split(/([\.])/).first.concat file_suffix
-  end
-
-  def file_suffix
-    "_#{self.class.name.downcase}"
-  end
-
-  def ident_r
-    Regexp.new "("+IDENTITIES.join("|")+")"
-  end
-
-  def units_r
-    Regexp.new( "("+UNITS.join("|")+")" ).freeze
-  end
 end
-
-
